Update OpenAPI schema transforms
